+++ RMS(Server)/RMS(Server)/Models/Inventory.cs
@@ -10,8 +10,8 @@ namespace RMS_Server_.Models
         public string Id { get; set; }
         public List<InventoryHistoryModel> InventoryHistoryModel { get; set; }
         public string Name { get; set; }
-        public int UsedQuantity { get; set; }
-        public int RemainingQuantity { get; set; }  
+        public float UsedQuantity { get; set; }
+        public float RemainingQuantity { get; set; }  
         public  string Unit { get; set; }
         public int AveragePrice { get; set; }
